+++ tests/pyutilkit/test_classes.py
@@ -1,0 +1,8 @@
+from pyutilkit.classes import Singleton
+
+
+def test_singleton() -> None:
+    class A(metaclass=Singleton):
+        pass
+
+    assert A() is A()