+++ docs/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog], and this project adheres to [Semantic
 -   A class to hold an SGR string
 -   A method to print a header spanning the whole terminal line
 -   A class to hold timings with ns resolution
+-   A metaclass for singletons
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html